% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{load_bertopic}
\alias{load_bertopic}
\title{Load a previously saved BERTopic model}
\usage{
load_bertopic(path)
}
\arguments{
\item{path}{Path to the directory created by \code{\link{save_bertopic}}.}
}
\value{
A \code{bertopic_fit} object.
}
\description{
Reads a model directory written by \code{\link{save_bertopic}} and
reconstructs the \code{bertopic_fit} object.
}
\seealso{
\code{\link{save_bertopic}}
}
