% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{print.bertopic_fit}
\alias{print.bertopic_fit}
\title{Print method for bertopic_fit objects}
\usage{
\method{print}{bertopic_fit}(x, ...)
}
\arguments{
\item{x}{A bertopic_fit object.}

\item{...}{Unused.}
}
\description{
Print method for bertopic_fit objects
}
