% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_flow.R
\name{print.bertopic_flow}
\alias{print.bertopic_flow}
\title{Print method for bertopic_flow objects}
\usage{
\method{print}{bertopic_flow}(x, ...)
}
\arguments{
\item{x}{A \code{bertopic_flow} object.}

\item{...}{Unused.}
}
\description{
Print method for bertopic_flow objects
}
