% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{print_topics}
\alias{print_topics}
\title{Pretty-print discovered topics}
\usage{
print_topics(fit, max_topics = 20)
}
\arguments{
\item{fit}{A fit object from \code{\link[=fit_bertopic]{fit_bertopic()}}.}

\item{max_topics}{Maximum number of topics to display.}
}
\description{
Pretty-print discovered topics
}
