% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{print.bert_encoder}
\alias{print.bert_encoder}
\title{Print method for bert_encoder objects}
\usage{
\method{print}{bert_encoder}(x, ...)
}
\arguments{
\item{x}{A bert_encoder object.}

\item{...}{Unused.}
}
\description{
Print method for bert_encoder objects
}
