% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{print.hf_classifier}
\alias{print.hf_classifier}
\title{Print method for hf_classifier objects}
\usage{
\method{print}{hf_classifier}(x, ...)
}
\arguments{
\item{x}{An hf_classifier object.}

\item{...}{Unused.}
}
\description{
Print method for hf_classifier objects
}
