% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_topics}
\alias{visualize_topics}
\title{Visualise documents in topic space}
\usage{
visualize_topics(
  fit,
  dims = NULL,
  label_topics = TRUE,
  n_label_words = 3L,
  point_size = 5L,
  noise_color = "#cccccc",
  width = 900L,
  height = 700L
)
}
\description{
Visualise documents in topic space
}
